# abc9: speedup by using using "clean" more efficiently